chore: renamed the customHeader -> Header and added headerButton to this module
