+++ app/containers/Header/components/HeaderButton/HeaderButtonItem.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Platform, StyleSheet, Text } from 'react-native';
 import { PlatformPressable } from '@react-navigation/elements';
 
-import { CustomIcon, TIconsName } from '../CustomIcon';
-import { useTheme } from '../../theme';
-import sharedStyles from '../../views/Styles';
+import { CustomIcon, TIconsName } from '../../../CustomIcon';
+import { useTheme } from '../../../../theme';
+import sharedStyles from '../../../../views/Styles';
 
 export interface IHeaderButtonItem {
 	title?: string;
