+++ app/containers/Header/components/HeaderTitle/styles.ts
