+++ app/views/E2EEnterYourPasswordView.tsx
@@ -5,7 +5,7 @@ import { useDispatch } from 'react-redux';
 
 import { encryptionDecodeKey } from '../actions/encryption';
 import Button from '../containers/Button';
-import * as HeaderButton from '../containers/HeaderButton';
+import * as HeaderButton from '../containers/Header/components/HeaderButton';
 import KeyboardView from '../containers/KeyboardView';
 import SafeAreaView from '../containers/SafeAreaView';
 import StatusBar from '../containers/StatusBar';
